# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fleet Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone number and PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /fuelings:
    get:
      operationId: listFuelings
      tags: [fuelings]
      summary: List all fuelings
      responses:
        "200":
          description: List of fuelings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fueling"
    post:
      operationId: createFueling
      tags: [fuelings]
      summary: Create fueling record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFuelingRequest"
      responses:
        "201":
          description: Fueling created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fueling"
  /fuelings/{id}:
    get:
      operationId: getFueling
      tags: [fuelings]
      summary: Get fueling by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fueling details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fueling"
    put:
      operationId: updateFueling
      tags: [fuelings]
      summary: Update fueling
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFuelingRequest"
      responses:
        "200":
          description: Fueling updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fueling"
    delete:
      operationId: deleteFueling
      tags: [fuelings]
      summary: Delete fueling
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /finance:
    get:
      operationId: listFinanceRecords
      tags: [finance]
      summary: List financial records
      responses:
        "200":
          description: List of financial records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinanceRecord"
    post:
      operationId: createFinanceRecord
      tags: [finance]
      summary: Create financial record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFinanceRecordRequest"
      responses:
        "201":
          description: Finance record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceRecord"
  /finance/{id}:
    get:
      operationId: getFinanceRecord
      tags: [finance]
      summary: Get finance record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Finance record details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceRecord"
    put:
      operationId: updateFinanceRecord
      tags: [finance]
      summary: Update finance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFinanceRecordRequest"
      responses:
        "200":
          description: Finance record updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceRecord"
    delete:
      operationId: deleteFinanceRecord
      tags: [finance]
      summary: Delete finance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /reports/fuelings:
    get:
      operationId: getFuelingsReport
      tags: [reports]
      summary: Get fueling report data
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Fueling report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FuelingsReport"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    LoginRequest:
      type: object
      properties:
        phone:
          type: string
        pin:
          type: string
      required: [phone, pin]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, driver]
        active:
          type: boolean
        vehicleId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, name, phone, role, active, createdAt]
...
    Fueling:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehiclePlate:
          type: string
          nullable: true
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
          nullable: true
        date:
          type: string
        liters:
          type: number
        pricePerLiter:
          type: number
        totalCost:
          type: number
        mileage:
          type: number
        station:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, vehicleId, date, liters, pricePerLiter, totalCost, mileage, createdAt]
    CreateFuelingRequest:
      type: object
      properties:
        vehicleId:
          type: integer
        driverId:
          type: integer
          nullable: true
        date:
          type: string
        liters:
          type: number
        pricePerLiter:
          type: number
        totalCost:
          type: number
        mileage:
          type: number
        station:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [vehicleId, date, liters, pricePerLiter, totalCost, mileage]
    UpdateFuelingRequest:
      type: object
      properties:
        vehicleId:
          type: integer
        driverId:
          type: integer
          nullable: true
        date:
          type: string
        liters:
          type: number
        pricePerLiter:
          type: number
        totalCost:
          type: number
        mileage:
          type: number
        station:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    FinanceRecord:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        date:
          type: string
        vehicleId:
          type: integer
          nullable: true
        vehiclePlate:
          type: string
          nullable: true
        referenceId:
          type: integer
          nullable: true
        referenceType:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, type, category, description, amount, date, createdAt]
    CreateFinanceRecordRequest:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        date:
          type: string
        vehicleId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [type, category, description, amount, date]
    UpdateFinanceRecordRequest:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        date:
          type: string
        vehicleId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
...
    FuelingsReport:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/Fueling"
        totalLiters:
          type: number
        totalCost:
          type: number
        averagePricePerLiter:
          type: number
        byVehicle:
          type: array
          items:
            type: object
            properties:
              vehiclePlate:
                type: string
              totalLiters:
                type: number
              totalCost:
                type: number
            required: [vehiclePlate, totalLiters, totalCost]
      required: [records, totalLiters, totalCost, averagePricePerLiter, byVehicle]
...